Change version check timing
