+++ scripts/release-master/is-release-ready.js
@@ -12,26 +12,6 @@ module.exports = async function isReleaseReady(version) {
   console.log('=> Yes. continue release steps');
   console.log('');
 
-  console.log(`Script tag for v${version} exists in each example file?`);
-  const cond2 = isAllExampleSdkVersionURLCorrect(version);
-  if (!cond2) {
-    console.log('=> No. abort release steps');
-    console.log('');
-    return false;
-  }
-  console.log('=> Yes. continue release steps');
-  console.log('');
-
-  console.log(`v${version} exists in README.md?`);
-  const cond3 = isReadmeVersionURLCorrect(version);
-  if (!cond3) {
-    console.log('=> No. abort release steps');
-    console.log('');
-    return false;
-  }
-  console.log('=> Yes. continue release steps');
-  console.log('');
-
   return true;
 };
 
@@ -53,31 +33,4 @@ async function hasChangeLog(version) {
 
     rl.once('close', () => resolve(false));
   });
-}
-
-function isSdkVersionURLCorrect(version, filepath) {
-  const data = fs.readFileSync(filepath, 'utf8');
-  const matches = data.match(
-    /cdn\.webrtc\.ecl\.ntt\.com\/skyway-([0-9]+\.[0-9]+\.[0-9]+)(\.min)?\.js/g
-  );
-  for (const match of matches) {
-    if (!match.includes(version)) return false;
-  }
-  return true;
-}
-
-function isAllExampleSdkVersionURLCorrect(version) {
-  const examplePaths = [
-    './examples/p2p-data/index.html',
-    './examples/p2p-media/index.html',
-    './examples/room/index.html',
-  ];
-  for (const path of examplePaths) {
-    if (!isSdkVersionURLCorrect(version, path)) return false;
-  }
-  return true;
-}
-
-function isReadmeVersionURLCorrect(version) {
-  return isSdkVersionURLCorrect(version, './README.md');
 }
